backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
psycopg2
python-dotenv
Pillow
gunicorn
dj-database-url
spacy
scikit-learn
pandas